nxp-ls1028: create graphic kernel config file for ls1028 platform
Issue: LINUXEXEC-19927

NXP ls1028 is a special SoC, because it is the only one supporting graphic
feature in Qoriq series SoC.  In additional, there is a limitation of using
GPU module, refer to lsdk 21.08, section 7.2.18, GPU driver does not
support SMMU feature, it needs to add iommu.passthrough=1 and
arm-smmu.disable_bypass=0 parameter in bootargs. Therefore, create
a graphic kernel config file to enable graphic feature explicitly. In this
way, SMMU functionality will not be affected if graphic feature is not
enabled.

Signed-off-by: Meng Li <Meng.Li@windriver.com>