nxp-ls1028: revert meaningless patch of nxp-ls1028 platform
Issue: LIN1022-1801

Add a patch to revert the patch that is not compatible with
nxp-ls1028 platform, because the libgbm is installed in usr/lib64
for nxp-ls1028 platform, and will be loaded by drm-backend.
The patch for imx93 platform is a specific one, and should not
be added to  the common code without any macro limitation.

Signed-off-by: Meng Li <Meng.Li@windriver.com>