nxp-ls1028: remove incorrect setting in mesa-demos bbappend
Issue: LIN1022-1822

In latest oe-core, there is a commit a904c9492e86("mesa-demos:
drop glx option and patch"), glx had been removed, so it doesn't
make sense to add glx config again. Moreover, the x11 config is
also not necessary. In meta-freescale layer, the x11 config has
been remove in  mesa-demos bbappend, so it doesn't need to remove
again in the bbapend of the nxp-ls1028 layer.

Signed-off-by: Meng Li <Meng.Li@windriver.com>