nxp-ls1028: improve wpa-supplicant bbappend that is from upstream meta-qoriq layer
Issue: LINUXEXEC-26882

Get wpa-supplicant bbappend from upstream meta-qoriq layer, and reuse
the defconfig in meta-imx/meta-bsp layer. Because the variable of
FILESEXTRAPATHS in meta-imx/meta-bsp is limited with OVERRIDES
imx-nxp-bsp, it doesn't exist in nxp-ls1028 BSP. So, LAYER_PATH_meta-bsp
is create in meta-imx/meta-bsp layer, so that defconfig is able to be
reused in nxp-ls1028 layer.

Signed-off-by: Meng Li <Meng.Li@windriver.com>