nxp-ls1028: get wpa-supplicant bbappend that is from upstream meta-qoriq layer
Issue: LINUXEXEC-31387

On ls1028 platfrom, the defconfig is not used. But in old meta-imx layer,
wpa-supplicant bbappend file installs the defconfig by defaut. So, it needs
to remove defconfig file based on old old meta-imx layer. Currently, in
latest meta-imx layer, the defconfig file installation is controlled by overrides
mx8-nxp-bsp, and defconfig will not be installed for ls1028 platform. So, it
doesn't need to remove it explicitly.

Signed-off-by: Meng Li <Meng.Li@windriver.com>